% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_MO.R
\name{MO}
\alias{MO}
\title{Compute the probability that a mutation ocurrs on branches of a phylogeny}
\usage{
MO(
  alpha,
  beta,
  lambda01,
  lambda02,
  lambda12,
  initial_obs,
  t,
  binary = c(TRUE, FALSE),
  multinomialPrior = c(FALSE, TRUE),
  seed = 1000
)
}
\arguments{
\item{alpha}{false positive error rate/probability}

\item{beta}{false negative error rate/probability}

\item{lambda01}{unit_theta=lambda01 rate from 0 to 1}

\item{lambda02}{unit_gamma=lambda02 rate from 0 to 2}

\item{lambda12}{unit_mu=lambda12 rate from 1 to 2}

\item{initial_obs}{observed mutation data matrix of all cells in the tree, each row is one mutation and each column is one cell}

\item{t}{phylogenetic tree}

\item{binary}{a TRUE/FALSE of whether the data is binary or ternary}

\item{multinomialPrior}{a TRUE/FALSE of whether the prior of mutations are multinomial}
}
\value{
A probability matrix of the mutations on each branch
}
\description{
This function computes the probability that a mutation occurs on each branch of a tree. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
